refactor: move role marker to the side
